feat: crud user

+++ src/modules/users/user.service.ts
@@ -2,43 +2,61 @@ import { User } from "@prisma/client";
 import { CreateUserDto } from "./dtos/createUser.dto.js";
 import UserRepository from "./user.repository.js";
 import { UpdateUserDto } from "./dtos/updateUser.dto.js";
+import { ResponseUserDto } from "./dtos/responseUser.dto.js";
 
 class UserService {
-    private readonly userRepository: UserRepository;
-    constructor() {
-        this.userRepository = new UserRepository();
+  private readonly userRepository: UserRepository;
+  private readonly userResponse: ResponseUserDto;
+
+  constructor() {
+    this.userRepository = new UserRepository();
+    this.userResponse = new ResponseUserDto();
+  }
+
+  public createUser = async (user: CreateUserDto): Promise<ResponseUserDto> => {
+    const userCreated = await this.userRepository.createUser(user);
+    return this._mapUserToResponse(userCreated);
+  };
+
+  public findAll = async (): Promise<ResponseUserDto[]> => {
+    const users = await this.userRepository.findAll();
+    return users.map((user) => this._mapUserToResponse(user));
+  };
+
+  public findById = async (id: string): Promise<ResponseUserDto> => {
+    const user = await this.userRepository.findById(id);
+
+    if (!user) {
+      throw new Error("User not found");
     }
 
-    async createUser(user: CreateUserDto): Promise<User> {
-        return await this.userRepository.createUser(user);
+    return this._mapUserToResponse(user);
+  };
+
+  async updateUser(user: UpdateUserDto): Promise<ResponseUserDto> {
+    const userOnDb = await this.findById(user.id);
+
+     if (!userOnDb) {
+       throw new Error("User not found");
+     }
+
+    const userUpdate = await this.userRepository.updateUser(user);
+
+    return this._mapUserToResponse(userUpdate);
+  }
+
+  async deleteUser(id: string): Promise<void> {
+    const userOnDb = await this.findById(id);
+
+    if (!userOnDb) {
+      throw new Error("User not found");
     }
+    await this.userRepository.deleteUser(id);
+  }
 
-    async findAll(): Promise<User[]> {
-        return await this.userRepository.findAll();
-    }
-
-    async findById(id: string): Promise<User> {
-        const user = await this.userRepository.findById(id);
-
-        if (!user) {
-            throw new Error("User not found");
-        }
-
-        return user;
-    }
-
-    async updateUser(user: UpdateUserDto): Promise<User> {
-        const userDb = await this.findById(user.id);
-
-        const userUpdate = await this.userRepository.updateUser(user);
-
-        return userUpdate;
-    }
-
-    async deleteUser(id: string) : Promise<void> {
-        const user = await this.findById(id);
-        await this.userRepository.deleteUser(id);
-    }
+  private _mapUserToResponse = (user: User): ResponseUserDto => {
+    return Object.assign(this.userResponse, user);
+  };
 }
 
 export default UserService;
